+++ xkcd/src/main/java/com/android/xkcd/fragment/AltTextDialog.java
@@ -1,10 +1,12 @@
-package com.android.xkcd;
+package com.android.xkcd.fragment;
 
 import android.support.v7.app.AlertDialog;
 import android.app.Dialog;
 import android.support.v4.app.DialogFragment;
 import android.content.DialogInterface;
 import android.os.Bundle;
+
+import com.android.xkcd.R;
 
 /**
  * Created by max on 14/07/17.
