+++ xkcd/src/main/res/layout/activity_main.xml
@@ -3,7 +3,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools" android:layout_width="match_parent"
-    android:layout_height="match_parent" tools:context="com.android.xkcd.MainActivity">
+    android:layout_height="match_parent" tools:context="com.android.xkcd.activity.MainActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
